Moveable Pieces added. (#2)
* Piece Moving now implemented in Project_main.py

Added the ability to move pieces on the checkerboard when you click on them. When a piece is clicked all potential moves around the piece will be shown and when you click on one of those potential pieces the piece will move there if it's a legal position. In order to make sure a piece isn't moved to an incorrect position, an is_legal function has been added.

* Added FPS along with new comments in settings.py.

- FPS is used to set the frames per second of the application so that pygame doesn't update far too often.
- Changed some colour settings and added some additional comments and removed incorrect ones.