Turns implemented (#5)
* Project_Main code moved to Project_GameState

Code from Project_Main moved to Project_GameState so we can do turns in the program and in the future more easily implement AI in the code.

* Project_Main code removed and turns implemented

Project_Main code was removed to be put in Project_Gamestate in order to implement turns and so AI is more easy to implement later.

* Bug Fixes.

red_piece_list are reinitialized each time the program starts. So the STARTING_RED/BLACK_POSITIONS array aren't linked together.

* Bugs fixes

I messed up the commit. Went to master by accident so check the commit for that for a description.